Rename plugin.py and move from plugins dir to plugins.py
