+++ pyomo/checker/hooks.py
@@ -11,7 +11,7 @@
 
 __all__ = ['IPreCheckHook', 'IPostCheckHook']
 
-from pyomo.common.plugin import Interface
+from pyomo.common.plugin_base import Interface
 
 class IPreCheckHook(Interface):
