+++ pyomo/checker/plugins/checkers/model/imports.py
@@ -10,7 +10,7 @@
 #  ___________________________________________________________________________
 
 import ast
-import pyomo.common.plugin
+import pyomo.common.plugin_base
 
 from pyomo.checker.plugins.checker import IterativeTreeChecker
 
@@ -21,7 +21,7 @@ class Imports(IterativeTreeChecker):
     exists somewhere within the initial imports block
     """
 
-    pyomo.common.plugin.alias('model.imports', 'Check if pyomo.core or pyomo.environ has been imported.')
+    pyomo.common.plugin_base.alias('model.imports', 'Check if pyomo.core or pyomo.environ has been imported.')
 
     def beginChecking(self, runner, script):
         self.pyomoImported = False
