+++ pyomo/checker/plugins/checkers/py3k/range.py
@@ -10,13 +10,13 @@
 #  ___________________________________________________________________________
 
 import ast
-import pyomo.common.plugin
+import pyomo.common.plugin_base
 from pyomo.checker.plugins.checker import IterativeTreeChecker
 
 
 class XRange(IterativeTreeChecker):
 
-    pyomo.common.plugin.alias('py3k.xrange', 'Check if the xrange() function is used.')
+    pyomo.common.plugin_base.alias('py3k.xrange', 'Check if the xrange() function is used.')
 
     def check(self, runner, script, info):
         if isinstance(info, ast.Name):
