+++ pyomo/common/plugin_base.py
