+++ pyomo/common/plugins.py
