+++ pyomo/common/tests/test_plugin.py
@@ -14,7 +14,7 @@ import weakref
 
 from pyomo.common.unittest import TestCase
 from pyomo.common.log import LoggingIntercept
-from pyomo.common.plugin import (
+from pyomo.common.plugin_base import (
     Interface, Plugin, SingletonPlugin, ExtensionPoint, implements, alias,
     PluginFactory, PluginError, PluginGlobals, DeprecatedInterface
 )
